+++ components/collector/src/model/__init__.py
@@ -1,7 +1,7 @@
 """Model package."""
 
 from .entity import Entities, Entity
-from .issue_status import Issue, IssueStatus, IssueStatusCategory, IssueSprint
+from .issue_status import Issue, IssueRelease, IssueStatus, IssueStatusCategory, IssueSprint
 from .measurement import SourceMeasurement, MetricMeasurement
 from .parameters import SourceParameters
 from .responses import SourceResponses