+++ docs/src/changelog.md
@@ -28,7 +28,7 @@ If your currently installed *Quality-time* version is not v4.1.0, please read th
 - Allow the configuration of multiple LDAP servers, see the [deployment instructions](https://quality-time.readthedocs.io/en/latest/deployment.html#ldap) for more information on configuring LDAP. Closes [#4141](https://github.com/ICTU/quality-time/issues/4141).
 - Mention the repository owner in the [README.md](https://github.com/ICTU/quality-time/blob/master/README.md) and add [contributing guidelines](https://github.com/ICTU/quality-time/blob/master/CONTRIBUTING.md) and a [code of conduct](https://github.com/ICTU/quality-time/blob/master/CODE_OF_CONDUCT.md) to make the *Quality-time* repository comply with the [ICTU GitHub policy](https://github.com/ICTU/github-policy). Closes [#4142](https://github.com/ICTU/quality-time/issues/4142).
 - Add an explanation of what version numbers are supported for the 'source version' and 'software version' metrics. Closes [#4146](https://github.com/ICTU/quality-time/issues/4146).
-- In the issue popups, and in the issue cards if so configured via the settings, show the due date and the sprint of issues if they have them. Closes [#4186](https://github.com/ICTU/quality-time/issues/4186).
+- In the issue popups, and in the issue cards if so configured via the settings, show the due date, the release, and the sprint of issues if they have them. Closes [#4186](https://github.com/ICTU/quality-time/issues/4186).
 
 ### Changed